requirements.txt
setuptools
pynput
lupa
pyclipper
pillow
pyyaml
flask
waitress
requests
ruamel.yaml
backports.functools_lru_cache==1.6.6 ; python_version < '3.0'
